pyproject.toml
[tool.poetry]
name = "authservice"
version = "0.1.3"
description = ""
authors = ["GabrielCarvalho <gabrielcarvalho.workk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
fastapi = {extras = ["all"], version = "^0.111.0"}
pydantic-settings = "^2.3.4"
python-jose = "^3.3.0"
bcrypt = "^4.1.3"
sqlalchemy = {extras = ["postgresql-asyncpg"], version = "^2.0.29"}
psycopg-binary = "^3.1.18"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
python-slugify = "^8.0.4"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"
taskipy = "^1.12.2"
ruff = "^0.1.8"
pytest-asyncio = "^0.23.2"
factory-boy = "^3.3.0"
freezegun = "^1.4.0"
faker = "^24.1.0"
ipykernel = "^6.29.3"
pytest = "^8.1.1"
pytest-postgresql = "^6.0.0"
icecream = "^2.1.3"
psycopg2-binary = "^2.9.10"
types-python-slugify = "^8.0.2.20240310"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = 'uvicorn app.main:app --reload'
lint = 'ruff .'
pre_test = 'task lint'
test = 'pytest -s -x --capture=no --cov=app -vv'
verbose_test = 'pytest --verbose --show-capture=all --exitfirst --cov=app --cov-report=term-missing -vv'
commit_hook = "pre-commit run --all-files"
post_verbose_test = 'coverage html'
post_test = 'coverage html'
pre_setup = 'docker run -d --name multi-tenancy -e POSTGRES_USER=app_user -e POSTGRES_PASSWORD=app_password -e POSTGRES_DB=app_db -p 5432:5432 postgres:14'
setup = 'alembic upgrade head'

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.8
target-version = "py38"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = ["E701"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
